Move static to root package
It no longer depends on UI things.
